requirements.txt
pandas~=1.5.3
requests~=2.31.0
django~=4.1
neomodel~=5.2.1
djangorestframework~=3.14.0
mongoengine~=0.27.0
pydantic~=1.10.12
luigi~=3.0.3
biopython~=1.79
cobra~=0.25.0
taxoniq~=1.0.1
django4-background-tasks~=1.2.9